Update dependencies and document storage retry fail
This is meant to ensure a single version of ms-* project is used in process.
The tests all pass, and the TypeScript build works, and when I run the program
against my dev subscription it succeeds. They use a tool called Rush in the
monorepo. I wonder if there is a problem keeping the version up-to-date in the
individual package.jsons.